file_type,permissionに関する処理をLongFormatクラスへ移管
