+++ 08.ls_object/ls.rb
@@ -1,4 +1,5 @@
 # frozen_string_literal: true
 
 require_relative 'ls_main'
+
 FileList.new.format_files
